@@ -2,12 +2,12 @@ import React from 'react';
 import { useParams } from 'react-router';
 import { Helmet } from 'react-helmet';
 import parse from 'html-react-parser';
-import SettingsContext from '../SettingsContext';
-import slide1 from '../resources/images/especialidad1.jpg';
-import slide2 from '../resources/images/especialidad2.jpg';
-import slide3 from '../resources/images/especialidad3.jpg';
-import slide4 from '../resources/images/especialidad4.jpg';
-import slide5 from '../resources/images/especialidad5.jpg';
+import SettingsContext, { Settings } from '../SettingsContext';
+import slide1 from '../resources/images/especialidad1.webp';
+import slide2 from '../resources/images/especialidad2.webp';
+import slide3 from '../resources/images/especialidad3.webp';
+import slide4 from '../resources/images/especialidad4.webp';
+import slide5 from '../resources/images/especialidad5.webp';
 import './Speciality.scss';
 
 const specials: Special[] = [
@@ -68,13 +68,9 @@ type Special = {
   caption: string;
 };
 
-interface MatchParams {
-  esp: string;
-}
-
 function Speciality(): JSX.Element {
-  const settings = React.useContext(SettingsContext);
-  const { esp } = useParams<MatchParams>();
+  const settings: Settings = React.useContext(SettingsContext);
+  const { esp } = useParams();
   const [data, setData] = React.useState<SpecialityState>({
     loaded: false,
   });
