@@ -1,12 +1,12 @@
 import React, { useRef } from 'react';
-import { RouteComponentProps } from 'react-router';
+import { useParams } from 'react-router';
 import { Modal, ModalBody, ModalFooter, Button } from 'reactstrap';
 import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import parse from 'html-react-parser';
 import BlogItem from '../../library/BlogItem';
 import './AppointmentManager.scss';
-import SettingsContext from '../../SettingsContext';
+import SettingsContext, { Settings } from '../../SettingsContext';
 import authService from '../api-authorization/AuthorizeService';
 import AddPostModal from './AddPostModal';
 import BlogPostResponse from '../../library/BlogPostResponse';
@@ -17,15 +17,8 @@ type PostManagerState = {
   page?: number;
 };
 
-interface MatchParams {
-  page: string;
-}
-
-type PostManagerProps = RouteComponentProps<MatchParams>;
-
-function PostManager(props: PostManagerProps): JSX.Element {
-  const { match } = props;
-  const params = match?.params ?? { page: 1 };
+function PostManager(): JSX.Element {
+  const { page } = useParams();
   const [addModal, setAddModal] = React.useState(false);
   const addToggle = () => setAddModal(!addModal);
   const [editModal, setEditModal] = React.useState(false);
@@ -36,17 +29,20 @@ function PostManager(props: PostManagerProps): JSX.Element {
   const alertToggle = () => setAlertModal(!alertModal);
   const [alertMessage, setAlertMessage] = React.useState<string>('');
   const [markedPost, setMarkedPost] = React.useState<BlogItem>();
-  const settings = React.useContext(SettingsContext);
+  const settings: Settings = React.useContext(SettingsContext);
   const [postData, setPostData] = React.useState<PostManagerState>({
     loaded: false,
     posts: null,
-    page: +params.page ?? 1,
+    page: page ? +page : 1,
   });
 
   const isFirstRun = useRef(true);
 
-  const fetchPosts = async (page: number, baseHref: string) => {
-    fetch(`${baseHref}api/blog/getallposts`)
+  const fetchPosts = async (pageToFetch: number, baseHref: string) => {
+    const token = await authService.getAccessToken();
+    fetch(`${baseHref}api/blog/getallposts`, {
+      headers: !token ? {} : { Authorization: `Bearer ${token}` },
+    })
       .then((response: { json: () => Promise<BlogPostResponse[]> }) =>
         response.json()
       )
@@ -81,14 +77,14 @@ function PostManager(props: PostManagerProps): JSX.Element {
         setPostData({
           loaded: true,
           posts: retrievedPosts,
-          page,
+          page: pageToFetch,
         });
       })
       .catch(() => {
         setPostData({
           loaded: false,
           posts: null,
-          page,
+          page: pageToFetch,
         });
       });
   };
@@ -141,23 +137,23 @@ function PostManager(props: PostManagerProps): JSX.Element {
 
   React.useEffect(() => {
     if (settings?.baseHref !== undefined) {
-      const page = +params.page || 1;
+      const currentPage = page ? +page : 1;
 
       if (isFirstRun.current) {
         isFirstRun.current = false;
 
-        fetchPosts(page, settings.baseHref);
+        fetchPosts(currentPage, settings.baseHref);
         return;
       }
 
-      setPostData({ loaded: false, page });
-      fetchPosts(page, settings.baseHref);
+      setPostData({ loaded: false, page: currentPage });
+      fetchPosts(currentPage, settings.baseHref);
     }
-  }, [match.params.page, params.page, settings?.baseHref]);
+  }, [page, settings?.baseHref]);
 
   if (postData.loaded) {
     return (
-      <div className="app app-home header-distance">
+      <div className="app app-home header-distance-l">
         <p>
           <br />
           <Button
