@@ -1,6 +1,6 @@
 /* eslint-disable react/prefer-stateless-function */
 import * as React from 'react';
-import { Route } from 'react-router';
+import { Route, Routes } from 'react-router';
 import Login from './Login';
 import Logout from './Logout';
 import ApplicationPaths, {
@@ -19,41 +19,41 @@ function logoutAction(name: string): JSX.Element {
 export default class ApiAuthorizationRoutes extends React.Component {
   render(): JSX.Element {
     return (
-      <>
-        <Route path="/" render={() => loginAction(LoginActions.Login)} />
+      <Routes>
+        <Route path="/" element={loginAction(LoginActions.Login)} />
         <Route
           path={ApplicationPaths.Login}
-          render={() => loginAction(LoginActions.Login)}
+          element={loginAction(LoginActions.Login)}
         />
         <Route
           path={ApplicationPaths.LoginFailed}
-          render={() => loginAction(LoginActions.LoginFailed)}
+          element={loginAction(LoginActions.LoginFailed)}
         />
         <Route
           path={ApplicationPaths.LoginCallback}
-          render={() => loginAction(LoginActions.LoginCallback)}
+          element={loginAction(LoginActions.LoginCallback)}
         />
         <Route
           path={ApplicationPaths.Profile}
-          render={() => loginAction(LoginActions.Profile)}
+          element={loginAction(LoginActions.Profile)}
         />
         <Route
           path={ApplicationPaths.Register}
-          render={() => loginAction(LoginActions.Register)}
+          element={loginAction(LoginActions.Register)}
         />
         <Route
           path={ApplicationPaths.LogOut}
-          render={() => logoutAction(LogoutActions.Logout)}
+          element={logoutAction(LogoutActions.Logout)}
         />
         <Route
           path={ApplicationPaths.LogOutCallback}
-          render={() => logoutAction(LogoutActions.LogoutCallback)}
+          element={logoutAction(LogoutActions.LogoutCallback)}
         />
         <Route
           path={ApplicationPaths.LoggedOut}
-          render={() => logoutAction(LogoutActions.LoggedOut)}
+          element={logoutAction(LogoutActions.LoggedOut)}
         />
-      </>
+      </Routes>
     );
   }
 }
